Change button size and element style
