+++ lib/routes/auto/fields/AutonomousRightRowNumberPad.dart
@@ -50,11 +50,12 @@ class _AutonomousRightrowNumberPadState
       children: [
         Center(
             child: SizedBox(
-          height: 40,
-          width: 150,
+          height: 50,
+          width: 200,
           child: TextField(
             controller: AutonomousValues.numberPadDisplayedValue,
             readOnly: true,
+            style: TextStyle(fontSize: 20),
             decoration: InputDecoration(
                 fillColor: Colors.white,
                 filled: true,
